@@ -148,7 +148,7 @@
     "@vitest/eslint-plugin": "^1.3.13",
     "cross-env": "10.0.0",
     "dts-for-context-bridge": "0.7.1",
-    "electron": "38.1.2",
+    "electron": "38.2.0",
     "electron-builder": "26.0.12",
     "electron-builder-notarize": "^1.5.2",
     "eslint": "^9.36.0",
@@ -168,7 +168,7 @@
     "husky": "^9.1.7",
     "inversify": "^7.10.1",
     "js-yaml": "^4.1.0",
-    "lint-staged": "^16.2.1",
+    "lint-staged": "^16.2.3",
     "minimist": "^1.2.8",
     "msw": "^2.11.2",
     "prettier": "^3.6.2",
