@@ -24,7 +24,7 @@
   "devDependencies": {
     "@playwright/test": "1.55.1",
     "@types/node": "^22",
-    "electron": "38.1.2",
+    "electron": "38.2.0",
     "typescript": "^5.9.2",
     "vitest": "^3.2.4"
   }
