@@ -46,11 +46,11 @@ export function WindowsDownloads(): JSX.Element {
           <h3 className="mt-0">Podman Desktop for Windows</h3>
           <div className="pt-8">
             <TelemetryLink
-              className="mt-auto no-underline hover:no-underline inline-flex text-white hover:text-white bg-purple-500 border-0 py-2 px-6 focus:outline-hidden hover:bg-purple-500 rounded-sm text-md font-semibold"
+              className="mt-auto no-underline hover:no-underline inline-flex text-white hover:text-white bg-purple-500 border-0 py-2 px-6 focus:outline-hidden hover:bg-purple-500 rounded-sm text-md font-semibold items-center"
               eventPath="download"
               eventTitle="download-windows"
               to={windows.setupX64}>
-              <FontAwesomeIcon size="1x" icon={faDownload} className="mr-2" />
+              <FontAwesomeIcon size="1x" icon={faDownload} className="mr-2 my-auto" />
               Download Now
             </TelemetryLink>
             <caption className="block w-full mt-1 text/50 dark:text-white/50">
